Remove unused method.
Not much reason to use a method here when you could just funcall the
method-entry's result-encoder slot. There isn't a good reason to specialize
method-entry that I can think of, and even if there were, it would be better
to have a generic function that actually did the encoding instead of
returning a thunk.