Handle non-call-specific errors and connection close in the acceptor.
EOF conditions here will always be errors reading from the socket (and not
from e.g. some file a method is trying to read) because EOF conditions
signalled in methods will be caught by the call processing function.

Other errors that occur here weren't part of processing a specific call,
and can't be returned to the caller. Since this might cause clients to hang
waiting for a result, these errors are basically fatal and we log them and
close the connection.