Abort-close the socket stream on error.
If an error has occurred, the socket can be various states, some of which
are not amenable to the normal flush-and-close process. Since we're just
dropping the connection, use an abort-close on the stream to prevent the
socket shutdown from trying to flush data later.

This fixes issues with worker hangs on windows when someone connects and
immediately send a TCP Reset: without this, an error is triggered, and the
attempt to close the socket normally hangs indefinitely.