Fix improper encoding function default.
The purposes of the :encoder argument to REGISTER-METHOD is to return a
surrogate value that will be properly encoded by JSON:ENCODE-JSON, not to
encode a value directly (moreover, JSON:ENCODE-JSON defaults to encoding
to *STANDARD-OUTPUT*, which doesn't do anybody any good). The default of
NIL is appropriate, because then the value will be used directly.